src/main/resources/cloud_functions/bucket_log/requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-bigquery-storage